pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comments-classifier"
version = "2.1.0"
description = "Ukrainian text comments classifier with lazy-loading optimization"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Volodymyr Dunaiev", email = "volmyrdunayev@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    # Core ML and Data Processing
    "pandas>=2.0.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "scikit-learn>=1.3.0",
    "datasets>=2.14.0",
    "evaluate>=0.4.0",
    "joblib>=1.3.0",
    "numpy>=1.24.0",
    
    # System Monitoring
    "psutil>=5.9.0",
    
    # UI Framework
    "streamlit>=1.28.0",
    
    # File and Configuration Handling
    "pyyaml>=6.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=0.19.0",
    
    # Type Hints (for older Python)
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    
    # Code Quality
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    
    # Pre-commit
    "pre-commit>=3.0.0",
]

[project.scripts]
comments-classifier = "src.ui.app_ui:main"

[project.urls]
Homepage = "https://github.com/yourusername/comments-classifier"
Repository = "https://github.com/yourusername/comments-classifier"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"]

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers"
pythonpath = "."
markers = [
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
